More refactoring for org.apache.myfaces.config.webparameters.MyfacesConfig
