+++ impl/src/main/resources/META-INF/WebConfigParamsLogger.vm
@@ -26,7 +26,7 @@ import java.util.logging.Logger;
 import jakarta.faces.application.ProjectStage;
 import jakarta.faces.context.FacesContext;
 
-import org.apache.myfaces.config.MyfacesConfig;
+import org.apache.myfaces.config.webparameters.MyfacesConfig;
 import org.apache.myfaces.util.WebConfigParamUtils;
 import org.apache.myfaces.util.lang.StringUtils;
